requirements.txt
oracledb>=1.1.0
python-dotenv>=1.0.0
requests>=2.28.0
sqlparse>=0.4.3
ujson>=5.8.0